Added Venue

+++ TDD/BlastAsia.DigiBook/BlastAsia.DigiBook.API/Controllers/AppointmentsController.cs
@@ -2,14 +2,16 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using BlastAsia.DigiBook.Domain.Appointments;
-using BlastAsia.DigiBook.Domain.Models.Appointments;
-using Microsoft.AspNetCore.JsonPatch;
 using BlastAsia.DigiBook.API.Utils;
+using BlastAsia.DigiBook.Domain.Appointments;
+using BlastAsia.DigiBook.Domain.Contacts;
+using BlastAsia.DigiBook.Domain.Employees;
+using BlastAsia.DigiBook.Domain.Models.Appointments;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.JsonPatch;
+using Microsoft.AspNetCore.Mvc;
 
-namespace BlastAsia.DigiBook.API.Controllers
+namespace BlastAsia.DigiBook.Api.Controllers
 {
     [Produces("application/json")]
     [Route("api/Appointments")]
@@ -18,11 +20,10 @@ namespace BlastAsia.DigiBook.API.Controllers
         private readonly IAppointmentService appointmentService;
         private readonly IAppointmentRepository appointmentRepository;
 
-        public AppointmentsController(IAppointmentService appointmentService,
-           IAppointmentRepository appointmentRepository)
+        public AppointmentsController(IAppointmentRepository appointmentRepository, IAppointmentService appointmentService)
         {
-            this.appointmentService = appointmentService;
             this.appointmentRepository = appointmentRepository;
+            this.appointmentService = appointmentService;
         }
 
         [HttpGet, ActionName("GetAppointments")]
@@ -31,6 +32,7 @@ namespace BlastAsia.DigiBook.API.Controllers
             var result = new List<Appointment>();
             if (id == null)
             {
+
                 result.AddRange(this.appointmentRepository.Retrieve());
             }
             else
@@ -42,41 +44,57 @@ namespace BlastAsia.DigiBook.API.Controllers
             return Ok(result);
         }
 
+        [HttpPost]
+        public IActionResult CreateAppointment(
+            [Bind] Appointment appointment)
+        {
+            try
+            {
+                if (appointment == null)
+                {
+                    return BadRequest();
+                }
+                var result = this.appointmentService.Save(Guid.Empty, appointment);
+
+                return CreatedAtAction("GetAppointments", new { id = appointment.AppointmentId }, result);
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e);
+            }
+        }
+
         [HttpDelete]
         public IActionResult DeleteAppointment(Guid id)
         {
+            var deletedAppointment = appointmentRepository.Retrieve(id);
+            if (deletedAppointment == null)
+            {
+                return NotFound();
+            }
             this.appointmentRepository.Delete(id);
 
             return Ok();
         }
 
-        [HttpPost]
-        public IActionResult CreateAppointment(
-           [FromBody]Appointment appointment)
-        {
-            var result = this.appointmentService.Save(Guid.Empty, appointment);
-
-            return CreatedAtAction("GetAppointments", new { id = result.AppointmentId }, result);
-        }
-
         [HttpPut]
         public IActionResult UpdateAppointment(
-            [FromBody] Appointment appointment, Guid id)
+            [FromBody] Appointment modifiedAppointment, Guid id)
         {
-            var existingAppointment = appointmentRepository.Retrieve(id);
-
-            existingAppointment.ApplyChanges(appointment);
-
-            this.appointmentService.Save(id, existingAppointment);
-
-
+            var appointment = appointmentRepository.Retrieve(id);
+            if (appointment == null)
+            {
+                return BadRequest();
+            }
+            appointment.ApplyChanges(modifiedAppointment);
+            appointmentService.Save(id, appointment);
             return Ok(appointment);
         }
+
         [HttpPatch]
-        public IActionResult PatchAppointment(
-            [FromBody]JsonPatchDocument PatchAppointment, Guid id)
+        public IActionResult PatchAppointment([FromBody]JsonPatchDocument patchedAppointment, Guid id)
         {
-            if (PatchAppointment == null)
+            if (patchedAppointment == null)
             {
                 return BadRequest();
             }
@@ -87,7 +105,7 @@ namespace BlastAsia.DigiBook.API.Controllers
                 return NotFound();
             }
 
-            PatchAppointment.ApplyTo(appointment);
+            patchedAppointment.ApplyTo(appointment);
             appointmentService.Save(id, appointment);
 
             return Ok(appointment);
