+++ TDD/BlastAsia.DigiBook/BlastAsia.DigiBook.API/Controllers/EmployeesController.cs
@@ -2,16 +2,15 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 using BlastAsia.DigiBook.Domain.Employees;
 using BlastAsia.DigiBook.Domain.Models.Employees;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.JsonPatch;
+using Microsoft.AspNetCore.Mvc;
 using BlastAsia.DigiBook.API.Utils;
 
-namespace BlastAsia.DigiBook.API.Controllers
+namespace BlastAsia.DigiBook.Api.Controllers
 {
-
     [Produces("application/json")]
     [Route("api/Employees")]
     public class EmployeesController : Controller
@@ -19,11 +18,12 @@ namespace BlastAsia.DigiBook.API.Controllers
         private readonly IEmployeeService employeeService;
         private readonly IEmployeeRepository employeeRepository;
 
-        public EmployeesController(IEmployeeService employeeService,
-           IEmployeeRepository employeeRepository)
+        public EmployeesController(
+            IEmployeeRepository employeeRepository
+            , IEmployeeService employeeService)
         {
-            this.employeeService = employeeService;
             this.employeeRepository = employeeRepository;
+            this.employeeService = employeeService;
         }
 
         [HttpGet, ActionName("GetEmployees")]
@@ -32,6 +32,7 @@ namespace BlastAsia.DigiBook.API.Controllers
             var result = new List<Employee>();
             if (id == null)
             {
+
                 result.AddRange(this.employeeRepository.Retrieve());
             }
             else
@@ -43,38 +44,54 @@ namespace BlastAsia.DigiBook.API.Controllers
             return Ok(result);
         }
 
+        [HttpPost]
+        public IActionResult CreateEmployee([FromBody] Employee employee)
+        {
+            try
+            {
+                if (employee == null)
+                {
+                    return BadRequest();
+                }
+                var result = this.employeeService.Save(Guid.Empty, employee);
+
+                return CreatedAtAction("GetEmployees", new { id = employee.EmployeeId }, result);
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e);
+            }
+        }
+
         [HttpDelete]
         public IActionResult DeleteEmployee(Guid id)
         {
+            var deletedEmployee = employeeRepository.Retrieve(id);
+            if (deletedEmployee == null)
+            {
+                return NotFound();
+            }
             this.employeeRepository.Delete(id);
 
             return Ok();
         }
 
-        [HttpPost]
-        public IActionResult CreateEmployee(
-           [FromBody]Employee employee)
-        {
-            var result = this.employeeService.Save(Guid.Empty, employee);
-
-            return CreatedAtAction("GetEmployees", new { id = result.EmployeeId }, result);
-        }
-
         [HttpPut]
         public IActionResult UpdateEmployee(
-            [FromBody] Employee employee, Guid id)
+            [FromBody] Employee modifiedEmployee, Guid id)
         {
-            var existingEmployee = employeeRepository.Retrieve(id);
-
-            existingEmployee.ApplyChanges(employee);
-
-            this.employeeService.Save(id, existingEmployee);
-
+            var employee = employeeRepository.Retrieve(id);
+            if (employee == null)
+            {
+                return BadRequest();
+            }
+            employee.ApplyChanges(modifiedEmployee);
+            employeeService.Save(id, employee);
             return Ok(employee);
         }
+
         [HttpPatch]
-        public IActionResult PatchEmployee(
-            [FromBody]JsonPatchDocument patchedEmployee, Guid id)
+        public IActionResult PatchEmployee([FromBody]JsonPatchDocument patchedEmployee, Guid id)
         {
             if (patchedEmployee == null)
             {
